chore: extract the cpi-caller from the memo program
