+++ programs/axelar-solana-its-cpi-caller/Cargo.toml
@@ -1,52 +1,53 @@
 [package]
-name = "axelar-solana-memo-program"
+name = "axelar-solana-its-cpi-caller"
 version = "0.1.0"
-description = "Utility program to test Axelar Gateway and other integrations"
+description = "CPI caller program for Axelar ITS integration"
 edition = "2021"
 
 [lib]
-name = "axelar_solana_memo_program"
+name = "axelar_solana_its_cpi_caller"
 crate-type = ["cdylib", "lib"]
 
 [features]
 no-entrypoint = []
 devnet-amplifier = [
     "axelar-solana-gateway/devnet-amplifier",
+    "axelar-solana-gas-service/devnet-amplifier", 
     "axelar-solana-its/devnet-amplifier",
     "program-utils/devnet-amplifier"
 ]
 stagenet = [
     "axelar-solana-gateway/stagenet",
-    "axelar-solana-its/stagenet",
+    "axelar-solana-gas-service/stagenet",
+    "axelar-solana-its/stagenet", 
     "program-utils/stagenet"
 ]
 testnet = [
     "axelar-solana-gateway/testnet",
+    "axelar-solana-gas-service/testnet",
     "axelar-solana-its/testnet",
     "program-utils/testnet"
 ]
 mainnet = [
     "axelar-solana-gateway/mainnet",
+    "axelar-solana-gas-service/mainnet",
     "axelar-solana-its/mainnet",
     "program-utils/mainnet"
 ]
 default = ["devnet-amplifier"]
 
 [dependencies]
-anchor-discriminators-macros.workspace = true
-axelar-solana-encoding.workspace = true
 axelar-solana-gateway = { workspace = true, features = ["no-entrypoint"] }
+axelar-solana-gas-service = { workspace = true, features = ["no-entrypoint"] }
 axelar-solana-its = { workspace = true, features = ["no-entrypoint"] }
 borsh.workspace = true
-hex.workspace = true
-mpl-token-metadata.workspace = true
 program-utils = { workspace = true, default-features = false }
 solana-program.workspace = true
+spl-associated-token-account = { workspace = true, features = ["no-entrypoint"] }
 
 [dev-dependencies]
 ethers-core.workspace = true
 evm-contracts-test-suite.workspace = true
-itertools.workspace = true
 rstest.workspace = true
 solana-program-test.workspace = true
 solana-sdk.workspace = true
